Login when needed and add sync methods
When creating the client we should not login right away. The reason
being that when initialiing the client for EF Core cannot be async.
The same reason exists for adding sync methods, as the value conversion
happening in EF Core isn't directly async, making it impossible to
safely use async methods